requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncmy==0.2.10
attrs==25.3.0
billiard==4.2.1
boto3==1.38.3
botocore==1.38.3
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Cython==3.0.12
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
fonttools==4.57.0
frozenlist==1.6.0
greenlet==3.1.1
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflect==7.5.0
jiter==0.9.0
jmespath==1.0.1
kombu==5.5.3
more-itertools==10.6.0
motor==3.7.0
multidict==6.1.0
numpy==2.2.5
openai==1.66.3
opencv-python==4.11.0.86
packaging==24.2
pika==1.3.2
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
PyMySQL==1.1.1
python-box==7.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
ratelimit==2.2.1
redis==4.2.0
requests==2.32.3
s3transfer==0.12.0
sanic==24.12.0
sanic-ext==24.12.0
sanic-routing==23.12.0
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.39
tenacity==9.0.0
tqdm==4.67.1
tracerite==1.1.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
